+++ TrackingAPIServices/Repository/ITrackingRepository.cs
@@ -26,7 +26,7 @@ namespace HeatMapAPIServices.Repository
         List<StatisticFunnel> getstatisticFunnelData(int trackingFunnelInfoID);
         object udpateFunnelTrackingInfo(udpateTrackingStepInfoRequest request, int v);
         object getStatisticHeatMap(int webID, int trackingInfoID, int from, int to, int device, int v);
-        Object getStatisticFunnel(int webID, int trackingFunnelInfoID, int from, int to, int userId);
+        Object getStatisticFunnel(int webID, int trackingFunnelInfoID, long from, long to, int userId);
         IActionResult udpateNameFunnelTrackingInfo(udpateTrackingNameInfoRequest request, int v);
     }
 }
